pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["Aditya Pratap Singh <adityapratapsingh51@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0"
python-dotenv = "^0.19.2"
django-environ = "^0.8.1"
django-cors-headers = "^3.10.1"
whitenoise = "^5.3.0"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
psycopg2 = "^2.9.3"
gunicorn = "^20.1.0"
PyJWT = "^2.3.0"
pycryptodomex = "^3.12.0"
cryptography = "^36.0.1"
google-auth = "^2.3.3"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
